+++ src/main/java/racingcar/domain/car/Cars.java
@@ -1,5 +1,6 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
+import racingcar.domain.result.Result;
 import racingcar.validator.CarsValidator;
 
 import java.util.*;
@@ -20,8 +21,12 @@ public class Cars {
     }
 
     public static Cars of(String inputCarNames) {
-        List<Name> carNames = Arrays.stream(inputCarNames.split(DELIMITER)).map(Name::valueOf).collect(Collectors.toList());
-        List<Car> cars = carNames.stream().map(Car::of).collect(Collectors.toList());
+        List<Name> carNames = Arrays.stream(inputCarNames.split(DELIMITER))
+                .map(Name::valueOf)
+                .collect(Collectors.toList());
+        List<Car> cars = carNames.stream()
+                .map(Car::of)
+                .collect(Collectors.toList());
         return new Cars(cars);
     }
