+++ src/test/java/racingcar/domain/CarTest.java
@@ -2,6 +2,9 @@ package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.car.Car;
+import racingcar.domain.car.Name;
+import racingcar.domain.car.Position;
 
 import static org.assertj.core.api.Assertions.assertThat;
